fix(build): compose the agent prompts with the wrappers
The boundary test reads the composed prompts from the build output, but the
plugin ran in the main build — which `pnpm validate` and CI both run *after*
`pnpm test`, so the test failed on a clean checkout. Move it next to the
wrappers, whose compiled scripts the sibling boundary tests already depend on
and which CI builds before testing, and emit into dist/bin so the main build's
existing dist/bin -> assets/bin copy still ships them.
