+++ docs/AGENTS.md
@@ -328,7 +328,7 @@ The server manager generates per-workspace config files (`codehydra-hooks.json`,
 
 Every agent session is given a shipped description of the environment it runs in: busy/idle semantics, worktree lifecycle ownership, the editor it can act on, and operator-driven workspace creation. It is not generated per workspace — `getSystemPromptPath()` on each server manager resolves it from the runtime dir (`runtimePath("bin/...")`, outside the ASAR) exactly like the hook handler.
 
-**Composition.** Sources live in `resources/prompts/`: `shared.md` plus a per-agent appendix. A plugin in `electron.vite.config.ts` concatenates them at build time into `assets/bin/codehydra-prompt-claude.md` and `codehydra-prompt-opencode.md`. Composition happens at build rather than launch because Claude accepts exactly one prompt source: `--append-system-prompt-file` is last-wins, and combining it with `--append-system-prompt` is rejected outright. Only Claude gets the `ch-bg` appendix — that wrapper is detectable only in Claude's `background_tasks`, and it is not on OpenCode's PATH.
+**Composition.** Sources live in `resources/prompts/`: `shared.md` plus a per-agent appendix. A plugin in `vite.config.bin.ts` concatenates them into `dist/bin/codehydra-prompt-claude.md` and `codehydra-prompt-opencode.md`, alongside the compiled wrappers, and the main build copies `dist/bin` into `assets/bin`. It runs with the wrappers rather than in the main build so that `pnpm build:wrappers` produces it — CI runs that before `pnpm test`, which is what lets a boundary test assert on the shipped artifacts. Composition happens at build rather than launch because Claude accepts exactly one prompt source: `--append-system-prompt-file` is last-wins, and combining it with `--append-system-prompt` is rejected outright. Only Claude gets the `ch-bg` appendix — that wrapper is detectable only in Claude's `background_tasks`, and it is not on OpenCode's PATH.
 
 **Delivery.** Claude: the provider passes `_CH_CLAUDE_SYSTEM_PROMPT` and `ch-claude` turns it into `--append-system-prompt-file`. The wrapper treats it as required — a missing variable fails startup like a missing settings path, because a session without it is a broken setup rather than a session with less context. OpenCode: the server manager puts the path in the `instructions` array of `OPENCODE_CONFIG_CONTENT`, which opencode merges last and renders into the system prompt as `Instructions from: <path>`. Note that opencode replaces arrays on merge (only `plugin` is unioned), so a user's own `instructions` entries do not survive.
