+++ src/modules/agent-module/codehydra-prompt.boundary.test.ts
@@ -1,12 +1,14 @@
 // @vitest-environment node
 /**
  * Boundary tests for the composed CodeHydra system prompts
- * (out/main/assets/bin/codehydra-prompt-{claude,opencode}.md).
+ * (dist/bin/codehydra-prompt-{claude,opencode}.md).
  *
- * The files are composed at build time from resources/prompts (shared + a
- * per-agent appendix) and passed to the agents by absolute path, so the
- * assertions target the built artifacts rather than the sources — that is what
- * actually ships, and it catches a composition that silently dropped a part.
+ * The files are composed by `pnpm build:wrappers` from resources/prompts
+ * (shared + a per-agent appendix) and passed to the agents by absolute path,
+ * so the assertions target the built artifacts rather than the sources — that
+ * is what actually ships, and it catches a composition that silently dropped a
+ * part. Same dependency as the wrapper boundary tests, which read the compiled
+ * scripts from the same directory.
  *
  * Asserting the load-bearing facts rather than the prose: rewording is free,
  * dropping a rule is not.
@@ -17,9 +19,9 @@ import { resolve, join } from "node:path";
 import { readFile, access } from "node:fs/promises";
 import { constants } from "node:fs";
 
-const ASSETS_BIN = resolve(__dirname, "../../../out/main/assets/bin");
-const CLAUDE_PROMPT = join(ASSETS_BIN, "codehydra-prompt-claude.md");
-const OPENCODE_PROMPT = join(ASSETS_BIN, "codehydra-prompt-opencode.md");
+const DIST_BIN = resolve(__dirname, "../../../dist/bin");
+const CLAUDE_PROMPT = join(DIST_BIN, "codehydra-prompt-claude.md");
+const OPENCODE_PROMPT = join(DIST_BIN, "codehydra-prompt-opencode.md");
 
 /** Budget agreed per agent file: additions have to displace something. */
 const MAX_WORDS = 300;
@@ -34,7 +36,7 @@ describe("composed agent system prompts", () => {
     for (const file of [CLAUDE_PROMPT, OPENCODE_PROMPT]) {
       await expect(
         access(file, constants.R_OK),
-        `${file} missing — run \`pnpm build\` to compose the prompts`
+        `${file} missing — run \`pnpm build:wrappers\` to compose the prompts`
       ).resolves.toBeUndefined();
     }
     claude = await readFile(CLAUDE_PROMPT, "utf-8");
